sync-agents/requirements.txt

pyodbc
requests
python-dotenv
pyinstaller
